+++ frontend/src/ts/modals/custom-text-popup.ts
@@ -5,10 +5,10 @@ import * as TestLogic from "../test/test-logic";
 import * as ChallengeController from "../controllers/challenge-controller";
 import Config, * as UpdateConfig from "../config";
 import * as Misc from "../utils/misc";
-import * as WordFilterPopup from "../modals/word-filter";
+import * as WordFilterPopup from "./word-filter";
 import * as Notifications from "../elements/notifications";
-import * as SavedTextsPopup from "../modals/saved-texts";
-import * as SaveCustomTextPopup from "../modals/save-custom-text";
+import * as SavedTextsPopup from "./saved-texts";
+import * as SaveCustomTextPopup from "./save-custom-text";
 import AnimatedModal from "../utils/animated-modal";
 import { text } from "stream/consumers";
